requirements.txt
asgiref==3.4.1
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.0
lxml==4.6.4
Pillow==8.4.0
psycopg-binary==3.0.4
psycopg2-binary==2.9.2
pytz==2021.3
spyne==2.13.16
sqlparse==0.4.2
